Re-extract fields to separate file
